feat: clear monaco utils
